feat(ref_parser): add git-based repo and branch detection utilities
- Introduced new `git` module in `ref_parser` for extracting repository info and current branch
- Support fetching origin remote URL, current branch, and parsing `FETCH_HEAD`, compatible with worktrees
- Moved existing git detection code from `mod.rs` to `git.rs` and re-exported functions
- Updated tests to reflect changes and improve PR lookup JSON construction

This enhances repository context detection by interfacing directly with git commands, improving robustness for fork detection and branch identification.
