+++ src/ref_parser/mod.rs
@@ -1,15 +1,19 @@
 //! Parse pull request and issue references into repository and number pairs, optionally including discussion comment IDs.
 
-use std::path::Path;
-use std::{fs, process::Command, sync::LazyLock};
+use std::sync::LazyLock;
 
 use regex::Regex;
 use url::Url;
 
 use crate::VkError;
 
+mod git;
 #[cfg(test)]
 mod tests;
+
+pub use git::{current_branch, repo_from_fetch_head, repo_from_origin};
+#[cfg(test)]
+pub(crate) use git::{current_branch_impl, repo_from_fetch_head_impl, repo_from_origin_impl};
 
 /// Fragment prefix for discussion comment IDs in GitHub URLs.
 const DISCUSSION_FRAGMENT: &str = "#discussion_r";
@@ -81,44 +85,6 @@ fn strip_git_suffix(name: &str) -> &str {
     name.strip_suffix(".git").unwrap_or(name)
 }
 
-/// Internal implementation of branch detection that accepts an optional directory.
-///
-/// When `dir` is `Some`, runs git in that directory; otherwise uses the current
-/// working directory.
-pub(crate) fn current_branch_impl(dir: Option<&Path>) -> Option<String> {
-    let mut cmd = Command::new("git");
-    cmd.args(["symbolic-ref", "--short", "HEAD"]);
-    if let Some(d) = dir {
-        cmd.current_dir(d);
-    }
-    let output = cmd.output().ok()?;
-    if !output.status.success() {
-        // Fails on detached HEAD or outside a git repo
-        return None;
-    }
-    let branch = String::from_utf8(output.stdout).ok()?;
-    Some(branch.trim().to_string())
-}
-
-/// Extract the current branch name using `git symbolic-ref`.
-///
-/// Uses `git symbolic-ref --short HEAD` to resolve the branch name, which
-/// works correctly with worktrees, linked gitdirs, and unborn branches where
-/// `.git` may be a file rather than a directory.
-///
-/// Returns `None` if not inside a Git repository, in detached HEAD state,
-/// or if the command fails.
-///
-/// # Examples
-///
-/// ```ignore
-/// // When on branch "feature-branch"
-/// assert_eq!(current_branch(), Some("feature-branch".to_string()));
-/// ```
-pub fn current_branch() -> Option<String> {
-    current_branch_impl(None)
-}
-
 fn parse_github_url(
     input: &str,
     resource: ResourceType,
@@ -181,81 +147,6 @@ pub fn parse_repo_str(repo: &str) -> Option<RepoInfo> {
     }
 }
 
-/// Internal implementation of `FETCH_HEAD` parsing that accepts an optional directory.
-///
-/// When `dir` is `Some`, runs git in that directory and resolves paths relative
-/// to it; otherwise uses the current working directory.
-pub(crate) fn repo_from_fetch_head_impl(dir: Option<&Path>) -> Option<RepoInfo> {
-    let mut cmd = Command::new("git");
-    cmd.args(["rev-parse", "--git-path", "FETCH_HEAD"]);
-    if let Some(d) = dir {
-        cmd.current_dir(d);
-    }
-    let output = cmd.output().ok()?;
-    if !output.status.success() {
-        return None;
-    }
-    let rel_path = String::from_utf8(output.stdout).ok()?;
-    // git rev-parse --git-path returns a path relative to the working directory
-    let full_path = dir.map_or_else(
-        || std::path::PathBuf::from(rel_path.trim()),
-        |d| d.join(rel_path.trim()),
-    );
-    let content = fs::read_to_string(full_path).ok()?;
-    content.lines().find_map(parse_repo_str)
-}
-
-/// Extract repository information from `FETCH_HEAD`.
-///
-/// Uses `git rev-parse --git-path FETCH_HEAD` to resolve the actual path to
-/// the `FETCH_HEAD` file, which works correctly with worktrees and linked
-/// gitdirs where `.git` may be a file rather than a directory.
-///
-/// Parses the first matching GitHub URL from the `FETCH_HEAD` file, which is
-/// written after `git fetch` operations.
-pub fn repo_from_fetch_head() -> Option<RepoInfo> {
-    repo_from_fetch_head_impl(None)
-}
-
-/// Internal implementation of origin remote URL parsing that accepts an optional directory.
-///
-/// When `dir` is `Some`, runs git in that directory; otherwise uses the current
-/// working directory.
-pub(crate) fn repo_from_origin_impl(dir: Option<&Path>) -> Option<RepoInfo> {
-    let mut cmd = Command::new("git");
-    cmd.args(["remote", "get-url", "origin"]);
-    if let Some(d) = dir {
-        cmd.current_dir(d);
-    }
-    let output = cmd.output().ok()?;
-    if !output.status.success() {
-        return None;
-    }
-    let url = String::from_utf8(output.stdout).ok()?;
-    parse_repo_str(url.trim())
-}
-
-/// Extract repository information from the `origin` remote URL.
-///
-/// Uses `git remote get-url origin` to retrieve the URL, which works correctly
-/// with worktrees and linked gitdirs. This identifies the user's fork when
-/// working on a forked repository.
-///
-/// Returns `None` if the `origin` remote is not configured or the URL cannot
-/// be parsed as a GitHub repository.
-///
-/// # Examples
-///
-/// ```ignore
-/// // When origin points to a GitHub repository
-/// let repo = repo_from_origin().expect("origin configured");
-/// assert_eq!(repo.owner, "fork-owner");
-/// assert_eq!(repo.name, "repo");
-/// ```
-pub fn repo_from_origin() -> Option<RepoInfo> {
-    repo_from_origin_impl(None)
-}
-
 fn parse_reference(
     input: &str,
     default_repo: DefaultRepo,
@@ -286,6 +177,19 @@ fn parse_reference(
 ///   or bare number (`42`).
 /// * `default_repo` - Optional `owner/repo` string for bare number resolution.
 ///
+/// # Examples
+///
+/// ```
+/// # use vk::ref_parser::parse_issue_reference;
+/// let (repo, number) = parse_issue_reference(
+///     "https://github.com/owner/repo/issues/42",
+///     None,
+/// ).expect("valid issue URL");
+/// assert_eq!(repo.owner, "owner");
+/// assert_eq!(repo.name, "repo");
+/// assert_eq!(number, 42);
+/// ```
+///
 /// # Errors
 ///
 /// Returns [`VkError::InvalidRef`] for malformed input,
@@ -309,6 +213,19 @@ pub fn parse_issue_reference<'a>(
 /// * `input` - PR reference: full URL (`https://github.com/o/r/pull/42`)
 ///   or bare number (`42`).
 /// * `default_repo` - Optional `owner/repo` string for bare number resolution.
+///
+/// # Examples
+///
+/// ```
+/// # use vk::ref_parser::parse_pr_reference;
+/// let (repo, number) = parse_pr_reference(
+///     "https://github.com/owner/repo/pull/42",
+///     None,
+/// ).expect("valid PR URL");
+/// assert_eq!(repo.owner, "owner");
+/// assert_eq!(repo.name, "repo");
+/// assert_eq!(number, 42);
+/// ```
 ///
 /// # Errors
 ///
